requirements-dev.txt
Django==1.11.4
djangorestframework==3.6.4
django-webpack-loader==0.5.0
django-cors-headers==2.1.0
django-rest-auth==0.9.1
django-allauth